Export webhook default's mutateSparkPodOption and its implementations in pkg/webhook so that external tools(e.g. kueue) can re-use the mutation logic.
